requirements.txt
click
label-studio-ml
rq
supervisor
uwsgi
google-cloud-storage